refactor help output around default and deploy commands
