+++ packages/cli/src/index.ts
@@ -54,6 +54,7 @@ import type { AuthConfig, GlobalConfig } from '@vercel-internals/types';
 import { VercelConfig } from '@vercel/client';
 import box from './util/output/box';
 import { ProxyAgent } from 'proxy-agent';
+import { help } from './args';
 
 const VERCEL_DIR = getGlobalPathConfig();
 const VERCEL_CONFIG_PATH = configFiles.getConfigFilePath();
@@ -152,6 +153,7 @@ const main = async () => {
   //  * a path to deploy (as in: `vercel path/`)
   //  * a subcommand (as in: `vercel ls`)
   const targetOrSubcommand = argv._[2];
+  const subSubCommand = argv._[3];
 
   // Currently no beta commands - add here as needed
   const betaCommands: string[] = ['rollback'];
@@ -171,6 +173,14 @@ const main = async () => {
   if (!targetOrSubcommand && argv['--version']) {
     console.log(pkg.version);
     return 0;
+  }
+
+  // Handle bare `-h` directly
+  const bareHelpOption = !targetOrSubcommand && argv['--help'];
+  const bareHelpSubcommand = targetOrSubcommand === 'help' && !subSubCommand;
+  if (bareHelpOption || bareHelpSubcommand) {
+    output.print(help());
+    return 2;
   }
 
   // Ensure that the Vercel global configuration directory exists
@@ -300,7 +310,7 @@ const main = async () => {
   }
 
   if (subcommand === 'help') {
-    subcommand = argv._[3] || 'deploy';
+    subcommand = subSubCommand || 'deploy';
     client.argv.push('-h');
   }
 
@@ -411,7 +421,7 @@ const main = async () => {
     targetCommand !== 'login' &&
     targetCommand !== 'dev' &&
     targetCommand !== 'build' &&
-    !(targetCommand === 'teams' && argv._[3] !== 'invite')
+    !(targetCommand === 'teams' && subSubCommand !== 'invite')
   ) {
     let user = null;
