+++ packages/cli/src/util/arg-common.ts
@@ -2,9 +2,6 @@ const ARG_COMMON = {
   '--help': Boolean,
   '-h': '--help',
 
-  '--platform-version': Number,
-  '-V': '--platform-version',
-
   '--debug': Boolean,
   '-d': '--debug',
 
@@ -28,6 +25,10 @@ const ARG_COMMON = {
   '--api': String,
 
   '--cwd': String,
+
+  // Deprecated
+  '--platform-version': Number,
+  '-V': '--platform-version',
 };
 
 export default () => ARG_COMMON;
